+++ src/app/page.tsx
@@ -1,33 +1,8 @@
-import Link from "next/link";
-import { compareDesc, format, parseISO } from "date-fns";
-import { allPosts, Post } from "contentlayer/generated";
-import { getMDXComponent } from "next-contentlayer/hooks";
+import { compareDesc } from "date-fns";
+import { allPosts } from "contentlayer/generated";
+import PostCard from "../components/PostCard";
 
-function PostCard(post: Post) {
-  const Content = getMDXComponent(post.body.code);
-
-  return (
-    <div className="mb-8 ">
-      <h2 className="text-xl ">
-        <Link
-          href={post.url}
-          className="text-blue-700 hover:text-blue-900"
-          legacyBehavior
-        >
-          {post.title}
-        </Link>
-      </h2>
-      <time dateTime={post.date} className="block mb-2 text-xs text-gray-600">
-        {format(parseISO(post.date), "LLLL d, yyyy")}
-      </time>
-      <div className="text-sm">
-        <Content />
-      </div>
-    </div>
-  );
-}
-
-export default function Home() {
+export default function HomePage() {
   const posts = allPosts.sort((a, b) =>
     compareDesc(new Date(a.date), new Date(b.date))
   );
