+++ src/components/PostCard.tsx
@@ -1,0 +1,30 @@
+import { Post } from "contentlayer/generated";
+import { getMDXComponent } from "next-contentlayer/hooks";
+import { format, parseISO } from "date-fns";
+import Link from "next/link";
+
+const PostCard = (post: Post) => {
+  const Content = getMDXComponent(post.body.code);
+
+  return (
+    <div className="mb-8 ">
+      <h2 className="text-xl ">
+        <Link
+          href={post.url}
+          className="text-blue-700 hover:text-blue-900"
+          legacyBehavior
+        >
+          {post.title}
+        </Link>
+      </h2>
+      <time dateTime={post.date} className="block mb-2 text-xs text-gray-600">
+        {format(parseISO(post.date), "LLLL d, yyyy")}
+      </time>
+      <div className="text-sm">
+        <Content />
+      </div>
+    </div>
+  );
+};
+
+export default PostCard;
